% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_tracks.R
\name{compare_tracks}
\alias{compare_tracks}
\title{Compare grid cell counts}
\usage{
compare_tracks(
  simulated_tracks,
  stations,
  land_barrier,
  visibility_graph,
  grid,
  multi_grid,
  snap_tolerance,
  cores = 2,
  ...
)
}
\arguments{
\item{simulated_tracks}{Tracks to which modeled tracks should be compared}

\item{stations}{A sf object comprised of receiver station locations with a buffer around them that represents the range of the receiver (polygons)}

\item{land_barrier}{A sf polygon of a barrier object around which tracks should be re-routed}

\item{visibility_graph}{A visibility graph created from the barrier object @seealso \link[pathroutr:prt_visgraph]{pathroutr::prt_visgraph}}

\item{grid}{When multi_grid = FALSE, a sf polygon grid; When multi_grid = TRUE, a list of sf polygon grids}

\item{multi_grid}{boolean}

\item{snap_tolerance}{The tolerance (in meters) at which an intersection between a station and a track should snap to the station centroid. It is recommended that the snap_tolerance be equal to the station buffer size.}

\item{cores}{Number of cores to use during paralllel processing. Defaults to 2.}
}
\value{
A data frame with counts and differences by grid cell ID ("gid")
}
\description{
Compare grid cell counts
}
