% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_grid}
\alias{create_grid}
\title{Create grid}
\usage{
create_grid(resolution, study_site, crs, type)
}
\arguments{
\item{resolution}{grid cell resolution in km. one-sided length of grid cell, assumes
desired grid cell is to be squared}

\item{study_site}{simple feature polygon object in which grid cells should be created}

\item{crs}{specification for desired coordinate reference system transformation}

\item{type}{\code{"polygons"} for grid cell polygons or \code{"centers"} for center
points of grid cells}
}
\value{
grid comprised of polygons or points depending on \code{type} parameter
}
\description{
This function creates a grid inside a specified polygon.
}
\examples{
# NOTE: The study site must be in a projected coordinate system (e.g. EPSG: 3857)
# when it is initially fed into the function create_grid

site_depth_3857 <- site_depth |> sf::st_transform(3857)
HS_100km_grid <- create_grid(100, site_depth_3857, 4269, "polygons")
}
