% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_deprecated.R
\name{cts}
\alias{cts}
\title{Grid counts}
\usage{
cts(sg, df)
}
\arguments{
\item{sg}{spatial grid created from \code{grid_res} function}

\item{df}{sf object of reconstructed, re-routed, optionally buffered tracks created from \code{sub_rrt} function}
}
\value{
A simple feature object with counts associated with grid cell IDs \code{"gid"}
}
\description{
The function allows you to calculate counts per grid cell in demonstrative modeled movement data.
}
