% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_shore_rcvs.R
\name{d_shore_rcvs}
\alias{d_shore_rcvs}
\title{Distance to shore and density of receivers}
\usage{
d_shore_rcvs(km, study_site, land_barrier, epsg)
}
\arguments{
\item{km}{grid cell resolution in km. one-sided length of grid cell, assumes desired grid cell is to be squared}

\item{study_site}{simple feature polygon object that encompasses the entire study site}

\item{land_barrier}{simple feature (multi)polygon object to route tracks around}

\item{epsg}{epsg code for desired coordinated system transformation}
}
\value{
A simple feature multipolygon object with information on distance to shore from grid cell center, receiver presence/absence, counts, and densities
}
\description{
This function determines the distance to shore and the density of receivers for each grid cell.
}
