% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth_cutoff.R
\name{depth_cutoff}
\alias{depth_cutoff}
\title{Different cut off}
\usage{
depth_cutoff(reconstructed, depth_limit)
}
\arguments{
\item{reconstructed}{data frame object consisting of results from iterative reconstruction process.
Needs columns \code{mean_depth} and \code{g_fit}; i.e., needs to have been passed through
\code{get_depth} and \code{gd_fts}.}

\item{depth_limit}{a depth value in meters that represents the new depth cut off of interest}
}
\value{
a data frame object containing the percentage of grid cells that contained
a good fit out of all grid cells that had an average depth (mean_depth) less
than or equal to the depth_limit
}
\description{
Different cut off
}
\examples{
# Apply depth_cutoff to list of grid resolutions

dif_depth <- lapply(res, depth_cutoff, depth_limit = 300) |>
  dplyr::bind_rows(.id = 'resolution') |>
  dplyr::mutate(resolution = as.numeric(resolution)) |>
  dplyr::left_join(dplyr::tibble(resolution = 1:4,
    res_name = c("100km", "50km", "25km", "10km")),
    by = "resolution") |>
  dplyr::mutate(
    res_name = ordered(res_name, levels = c("100km", "50km", "25km", "10km"))
  )
}
