% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_to_shore.R
\name{distance_to_shore}
\alias{distance_to_shore}
\title{Distance to shore and density of receivers}
\usage{
distance_to_shore(resolution, study_site, land_barrier, crs, stations)
}
\arguments{
\item{resolution}{grid cell resolution in km. one-sided length of grid cell, assumes desired grid cell is to be squared}

\item{study_site}{simple feature polygon object that encompasses the entire study site}

\item{land_barrier}{simple feature (multi)polygon object to route tracks around}

\item{crs}{coordinate reference system specification for transformation}

\item{stations}{a simple feature (multi)point object representing receiver locations with detection range buffer}
}
\value{
A simple feature multipolygon object with information on distance to shore from grid cell center, receiver presence/absence, counts, and densities
}
\description{
This function determines the distance to shore and the density of receivers for each grid cell.
}
