% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_deprecated.R
\name{.check_deprecated_dots}
\alias{.check_deprecated_dots}
\title{Check and Remap Deprecated Arguments from Dots}
\usage{
.check_deprecated_dots(
  dots,
  mapping = .DEPRECATED_ARG_MAP,
  env = parent.frame()
)
}
\arguments{
\item{dots}{A named list of arguments captured via \code{list(...)}.}

\item{mapping}{A named character vector. Defaults to the global master map.}

\item{env}{The environment of the calling function, defaulting to the parent
frame. Mostly used for testing.}
}
\description{
Check and Remap Deprecated Arguments from Dots
}
\keyword{internal}
