% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fo_land_barrier}
\alias{fo_land_barrier}
\title{American coastal Atlantic states polygon}
\format{
An sf object with 435 rows and three columns:
\describe{
\item{poly_uid}{Polygon ID, ranging from 1 to 434}
\item{line_uid}{Values of 1 (States) or 2 (Washington, DC)}
\item{geometry}{Well-known text POLYGON of the American Atlantic coastal shelf
in WGS84/Pseudo-Mercator, EPSG:3857}
}
}
\usage{
fo_land_barrier
}
\description{
American coastal Atlantic states polygon
}
\keyword{datasets}
