% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaps_filled.R
\name{gaps_filled}
\alias{gaps_filled}
\title{Gaps filled}
\usage{
gaps_filled(reconstructed)
}
\arguments{
\item{reconstructed}{data frame object consisting of results from iterative reconstruction process}
}
\value{
a data frame object with a percentage of grid cells that contained a
good fit out of those that lacked receivers
}
\description{
This function calculates how well the reconstructions closed gaps in the network
receiver array
}
\examples{
# Apply gaps_filled to list of grid resolutions

clsd_gps <- lapply(res, gaps_filled) |>
  dplyr::bind_rows(.id = 'resolution') |>
    dplyr::left_join(
      data.frame(
        resolution = as.character(1:4),
        res_name = c("100km", "50km", "25km", "10km")
      ),
      by = "resolution") |>
  dplyr::mutate(
    res_name = ordered(res_name, levels = c("100km", "50km", "25km", "10km"))
  )
}
