% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_depth.R
\name{get_depth}
\alias{get_depth}
\title{Calculate mean depth}
\usage{
get_depth(HSgrid, depth)
}
\arguments{
\item{HSgrid}{list object of multiple grids}

\item{depth}{simple feature point object with \code{"altitude"} attribute data}
}
\value{
A simple feature multipolygon object with mean depth per grid cell
}
\description{
This function allows you to calculate mean depth per grid cell.
}
