% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_counts.R
\name{grid_counts}
\alias{grid_counts}
\title{Grid counts}
\usage{
grid_counts(grid, rerouted_tracks)
}
\arguments{
\item{grid}{spatial grid created from \code{grid_res} function}

\item{rerouted_tracks}{sf object of reconstructed, re-routed, optionally buffered tracks created from \code{sub_rrt} function}
}
\value{
A simple feature object with counts associated with grid cell IDs \code{"gid"}
}
\description{
The function allows you to calculate counts per grid cell in demonstrative modeled movement data.
}
