% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_res.R
\name{grid_res}
\alias{grid_res}
\title{Create grid}
\usage{
grid_res(km, study_site, epsg, what = c("polygons", "centers"), simplify = 0.2)
}
\arguments{
\item{km}{grid cell resolution in km. one-sided length of grid cell, assumes
desired grid cell is to be squared}

\item{study_site}{simple feature polygon object in which grid cells should be created}

\item{epsg}{epsg numeric code for desired coordinate system transformation}

\item{what}{\code{"polygons"} for grid cell polygons or \code{"centers"} for center
points of grid cells}

\item{simplify}{numeric. By what proportion of \code{km} do you wish to simplify
the \code{study_site} polygon? Higher numbers give faster run times at the expense
of resolution. Defaults to 0.2.}
}
\value{
grid comprised of polygons or points depending on \code{what} parameter
}
\description{
This function creates a grid inside a specified polygon.
}
\examples{
# NOTE: The study site must be in a projected coordinate system (e.g. EPSG: 3857)
# when it is initially fed into the function grid_res

site_depth_3857 <- site_depth |> sf::st_transform(3857)
HS_100km_grid <- grid_res(100, site_depth_3857, 4269, "polygons")
}
