% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis.R
\name{power_analysis}
\alias{power_analysis}
\title{Run a power analysis}
\usage{
power_analysis(results, significance_level, power, delta, n = NULL)
}
\arguments{
\item{results}{the resulting data frame from the iterative methods process}

\item{significance_level}{numeric. the desired level of significance to achieve}

\item{power}{the desired level of power to achieve}

\item{delta}{numeric. the desired effect size to achieve}

\item{n}{integer. the number of replicates/sample size. should be assigned
\code{NULL} if desiring sample size}
}
\value{
A data frame object
}
\description{
This is a wrapper function for running a power analysis on iterative simulation
and reconstruction methods process.
}
\examples{
# Use power_analysis wrapper function on example results
# Grab 100km resolution
res <- lapply(sporeg::results, `[[`, 1) |>
  dplyr::bind_rows() |>
  # Calculate standard deviation by group
  dplyr::group_by(gid) |>
  dplyr::summarise(sd = sd(dif)) |>
  dplyr::ungroup()

anims <- 30
power <- 0.8
delta <- anims * 0.01 # a delta within 1\% of the total number of animals
sig.level <- 0.95

power_analysis(res, sig.level, power, delta)
}
