% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcv_chr.R
\name{rcv_chr}
\alias{rcv_chr}
\title{Receiver characteristics}
\usage{
rcv_chr(df)
}
\arguments{
\item{df}{data frame object consisting of results from iterative reconstruction process}
}
\value{
a data frame object with summary statistics that provide insight into
the locations of the receivers in the network receiver array
}
\description{
Receiver characteristics
}
\examples{
# Apply rcv_chr to list of grid resolutions

rcv_chr(res[[1]])
}
