% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/receiver_summary.R
\name{receiver_summary}
\alias{receiver_summary}
\title{Receiver characteristics}
\usage{
receiver_summary(reconstructed)
}
\arguments{
\item{reconstructed}{data frame object consisting of results from iterative reconstruction process}
}
\value{
a data frame object with summary statistics that provide insight into
the locations of the receivers in the network receiver array
}
\description{
Receiver characteristics
}
\examples{
# Apply rcv_chr to list of grid resolutions

receiver_summary(res[[1]])
}
