% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reroute_tracks.R
\name{reroute_tracks}
\alias{reroute_tracks}
\title{Re-route tracks around a barrier}
\usage{
reroute_tracks(track_data, crs, barrier, visibility_graph, buffer)
}
\arguments{
\item{track_data}{point location data with latitude and longitude information}

\item{crs}{epsg code for desired coordinate system transformation}

\item{barrier}{polygon or multipolygon sf object}

\item{visibility_graph}{visibility_graph object created from \code{pathroutr::prt_vis_graph} function}

\item{buffer}{desired buffer size for re-routed tracks; generally should
relate to range of receivers; distance should use the same units as the
final coordinate system}
}
\value{
A simple feature (multi)polygon object
}
\description{
This function allows you to re-route demonstrative movement data that has been
reconstructed using a movement model around a polygon barrier.
}
\examples{
crs <- 3857
barrier <- atlcoast
visibility_graph <- pathroutr::prt_visgraph(barrier)
buffer <- 650
track_data <- sporeg::subset |>
  dplyr::filter(ID \%in\% c("A69-9001-23338", "A69-9001-25494"))

tbuff650 <- reroute_tracks(track_data, crs, barrier, visibility_graph, buffer)

}
