% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{silence_doFuture_note}
\alias{silence_doFuture_note}
\title{Need to have doFuture in Imports due to momentuHMM, but we never actually call it.
This causes a NOTE in R CMD check}
\usage{
silence_doFuture_note()
}
\description{
Need to have doFuture in Imports due to momentuHMM, but we never actually call it.
This causes a NOTE in R CMD check
}
