% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_tracks.R
\name{simulate_tracks}
\alias{simulate_tracks}
\title{Simulate tracks}
\usage{
simulate_tracks(
  n_animals,
  study_site,
  theta,
  min_velocity,
  max_velocity,
  release_site,
  crs,
  n_days,
  initial_heading
)
}
\arguments{
\item{n_animals}{integer. quantity of desired animals to simulate}

\item{study_site}{simple feature polygon object that encompasses area where
tracks are allowed to be simulated.}

\item{theta}{argument from \code{glatos::crw_in_polygon} function}

\item{min_velocity}{numeric. minimum velocity from which to sample step length}

\item{max_velocity}{numeric. maximum velocity from which to sample step length}

\item{release_site}{simple feature polygon object in which simulated animals are
"released" or where simulated tracks begin}

\item{crs}{EPSG code for study_site polygon. Must be projected coordinate system}

\item{n_days}{integer. number of days that tracks should be simulated}

\item{initial_heading}{argument from \code{glatos::crw_in_polygon} function}
}
\value{
simple feature (multi)linestring object that represents individual simulated tracks
}
\description{
This function simulates tracks inside a specified polygon.
}
\examples{

n_animals <- 2
study_site <- fo_study_site
theta <- c(0, 1.74)
min_velocity <- 0.98
max_velocity <- 1.58
release_site <- fo_rel_site
crs <- 3857
n_days <- 30
initial_heading <- 0

tracks <- simulate_tracks(n_animals, study_site, theta, min_velocity, max_velocity, release_site, crs, n_days, initial_heading)
}
