% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_gaps.R
\name{summarize_gaps}
\alias{summarize_gaps}
\title{Gaps characteristics}
\usage{
summarize_gaps(reconstructed)
}
\arguments{
\item{reconstructed}{data frame object consisting of results from iterative reconstruction process}
}
\value{
a data frame object with summary statistics that provide insight into
the locations of the gaps in the network receiver array that were closed by
the reconstruction process
}
\description{
Gaps characteristics
}
\examples{
# Apply summarize_gaps to list of grid resolutions

clsd_gp_chars <- lapply(res, summarize_gaps) |>
  dplyr::bind_rows(.id = 'resolution') |>
    dplyr::left_join(
      data.frame(
        resolution = as.character(1:4),
        res_name = c("100km", "50km", "25km", "10km")
      ),
      by = "resolution") |>
  dplyr::mutate(
    res_name = ordered(res_name, levels = c("100km", "50km", "25km", "10km"))
  )
}
