% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_good_fits.R
\name{summarize_good_fits}
\alias{summarize_good_fits}
\title{Good fits summary statistics}
\usage{
summarize_good_fits(reconstructed)
}
\arguments{
\item{reconstructed}{data frame object consisting of results from iterative reconstruction process}
}
\value{
a data frame object with summary statistics from grid cells that contained good fits
}
\description{
This function summarizes minimum, maximum, and mean statistics for depth (mean_depth), distance from shore (d_shore), receiver count (count), receiver presence/absence (p_a), and the percentage of grid cells missing receivers for all of the good fits (g_fit).
}
\examples{
# Apply gd_fts to list of grid resolutions

gd_fit_char <- lapply(res, summarize_good_fits) |>
  dplyr::bind_rows(.id = 'resolution') |>
  dplyr::mutate(resolution = as.numeric(resolution)) |>
  dplyr::left_join(dplyr::tibble(resolution = 1:4,
    res_name = c("100km", "50km", "25km", "10km")),
    by = "resolution") |>
  dplyr::mutate(
    res_name = ordered(res_name, levels = c("100km", "50km", "25km", "10km"))
  )
}
