% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_deprecated.R
\name{warn_deprecated}
\alias{warn_deprecated}
\title{Utility function to warn about function deprecation}
\usage{
warn_deprecated(old_name, new_name)
}
\arguments{
\item{old_name}{old function name}

\item{new_name}{new function name}
}
\description{
Utility function to warn about function deprecation
}
\keyword{internal}
