% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cts.R
\name{cts}
\alias{cts}
\title{Counts}
\usage{
cts(sg, df)
}
\arguments{
\item{sg}{spatial grid created from \code{grid_res} function}

\item{df}{sf object of reconstructed, re-routed, optionally buffered tracks created from \code{sub_rrt} function}
}
\value{
A simple feature object with counts associated with grid cell IDs \code{"gid"}
}
\description{
The function allows you to calculate counts per grid cell in demonstrative modeled movement data.
}
\examples{
sg <- grid_res(50, study_site, epsg = 3857, what = "polygons")
subset_cts <- cts(sg, tbuff650)
# Save counts to shapfile
sf::st_write(subset_cts, dsn = "subset_cts08312023.shp", delete_dsn = TRUE)
}
