% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/den_rcvs.R
\name{den_rcvs}
\alias{den_rcvs}
\title{Summarize receiver density and counts}
\usage{
den_rcvs(gridded_receivers)
}
\arguments{
\item{gridded_receivers}{A simple feature polygon object (a grid) which has columns of \code{den_rcs} and \code{count}.
This is usually the result of \code{res_grid} which has been passed through \code{d_shore_rcvs};
i.e.: \code{res_grid() |> d_shore_rcvs()}.}
}
\value{
A data frame object with minimum, mean, and maximum of receiver densities and counts in km^-2
}
\description{
This function provides summary statistics on the densities and counts of receivers in a grid.
}
\examples{

den_rcvs(res[[1]])
}
