% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/den_rcvs.R
\name{den_rcvs}
\alias{den_rcvs}
\title{Summarize receiver density and counts}
\usage{
den_rcvs(df)
}
\arguments{
\item{df}{A simple feature polygon object (a grid)}
}
\value{
A data frame object with minimum, mean, and maximum of receiver densities and counts in km^-2
}
\description{
This function provides summary statistics on the densities and counts of receivers in a grid.
}
\examples{
# Apply den_rcvs to list of grid resolutions

library(sporeg)
library(dplyr)
library(sf)
library(data.table)
load(system.file("extdata", "res.Rda", package = "sporeg"))

rcv_dens <- lapply(res, den_rcvs)

rcv_dens <- data.table::rbindlist(rcv_dens, idcol = 'resolution') \%>\%
left_join(tibble(resolution = 1:4,
res_name = c("100km", "50km", "25km", "10km")),
by = "resolution") \%>\%
mutate(res_name = ordered(res_name, levels = c("100km", "50km", "25km", "10km")))
}
