% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_deprecated.R
\name{comp_trks}
\alias{comp_trks}
\alias{cts}
\alias{d_shore_rcvs}
\alias{den_rcvs}
\alias{dif_co}
\alias{gd_fts}
\alias{get_or}
\alias{gps_chr}
\alias{gps_fld}
\alias{grid_res}
\alias{powr}
\alias{rcv_chr}
\alias{simul_trks}
\alias{sub_rrt}
\alias{zero_var}
\title{Deprecated functions}
\usage{
comp_trks(
  sim_trks,
  stations,
  land_barrier,
  vis_graph,
  HSgrid,
  multi.grid,
  snap_tolerance,
  cores = 2,
  ...
)

cts(sg, df)

d_shore_rcvs(km, study_site, land_barrier, epsg, sts_pts)

den_rcvs(df)

dif_co(df, depth_limit)

gd_fts(df)

get_or(model)

gps_chr(df)

gps_fld(df)

grid_res(km, study_site, epsg, what)

powr(output, sig.level, power, delta, n = NULL)

rcv_chr(df)

simul_trks(
  anims,
  study_site,
  theta,
  vmin,
  vmax,
  rel_site,
  crs,
  n_days,
  initHeading
)

sub_rrt(track_data, CRS, barrier, vis_graph, buffer)

zero_var(df)
}
\arguments{
\item{sim_trks}{Tracks to which modeled tracks should be compared}

\item{stations}{A sf object comprised of receiver station locations with a buffer around them that represents the range of the receiver (polygons)}

\item{land_barrier}{A sf polygon of a barrier object around which tracks should be re-routed}

\item{vis_graph}{vis_graph object created from \code{pathroutr::prt_vis_graph} function}

\item{HSgrid}{When multi_grid = FALSE, a sf polygon grid; When multi_grid = TRUE, a list of sf polygon grids}

\item{multi.grid}{boolean}

\item{snap_tolerance}{The tolerance (in meters) at which an intersection between a station and a track should snap to the station centroid. It is recommended that the snap_tolerance be equal to the station buffer size.}

\item{cores}{Number of cores to use during paralllel processing. Defaults to 2.}

\item{sg}{spatial grid created from \code{grid_res} function}

\item{df}{data frame object consisting of results from iterative reconstruction process}

\item{km}{grid cell resolution in km. one-sided length of grid cell, assumes
desired grid cell is to be squared}

\item{study_site}{simple feature polygon object that encompasses the entire study site}

\item{epsg}{specification for desired coordinate reference system transformation}

\item{sts_pts}{a simple feature (multi)point object representing receiver locations with detection range buffer}

\item{depth_limit}{a depth value in meters that represents the new depth cut off of interest}

\item{model}{A model object created from the \code{nlme} package}

\item{what}{\code{"polygons"} for grid cell polygons or \code{"centers"} for center
points of grid cells}

\item{output}{the resulting data frame from the iterative methods process}

\item{sig.level}{numeric. the desired level of significance to achieve}

\item{power}{the desired level of power to achieve}

\item{delta}{numeric. the desired effect size to achieve}

\item{n}{integer. the number of replicates/sample size. should be assigned
\code{NULL} if desiring sample size}

\item{anims}{integer. quantity of desired animals to simulate}

\item{theta}{argument from \code{glatos::crw_in_polygon} function}

\item{vmin}{numeric. minimum velocity from which to sample step length}

\item{vmax}{numeric. maximum velocity from which to sample step length}

\item{rel_site}{simple feature polygon object in which simulated animals are}

\item{crs}{coordinate reference system specification for transformation}

\item{n_days}{integer. number of days that tracks should be simulated}

\item{initHeading}{argument from \code{glatos::crw_in_polygon} function}

\item{track_data}{point location data with latitude and longitude information}

\item{CRS}{epsg code for desired coordinate system transformation}

\item{barrier}{polygon or multipolygon sf object}

\item{buffer}{desired buffer size for re-routed tracks; generally should
relate to range of receivers; distance should use the same units as the
final coordinate system}
}
\description{
These functions are provided for backwards compatibility and will be removed
in a future release.
}
