% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_deprecated.R
\name{comp_trks}
\alias{comp_trks}
\alias{cts}
\alias{d_shore_rcvs}
\alias{den_rcvs}
\alias{dif_co}
\alias{gd_fts}
\alias{get_or}
\alias{gps_chr}
\alias{gps_fld}
\title{Deprecated functions}
\usage{
comp_trks(
  sim_trks,
  stations,
  land_barrier,
  vis_graph,
  HSgrid,
  multi.grid,
  snap_tolerance,
  cores = 2,
  ...
)

cts(sg, df)

d_shore_rcvs(km, study_site, land_barrier, epsg, sts_pts)

den_rcvs(df)

dif_co(df, depth_limit)

gd_fts(df)

get_or(model)

gps_chr(df)

gps_fld(df)
}
\arguments{
\item{sim_trks}{Tracks to which modeled tracks should be compared}

\item{stations}{A sf object comprised of receiver station locations with a buffer around them that represents the range of the receiver (polygons)}

\item{land_barrier}{A sf polygon of a barrier object around which tracks should be re-routed}

\item{vis_graph}{A visibility graph created from the barrier object @seealso \link[pathroutr:prt_visgraph]{pathroutr::prt_visgraph}}

\item{HSgrid}{When multi_grid = FALSE, a sf polygon grid; When multi_grid = TRUE, a list of sf polygon grids}

\item{multi.grid}{boolean}

\item{snap_tolerance}{The tolerance (in meters) at which an intersection between a station and a track should snap to the station centroid. It is recommended that the snap_tolerance be equal to the station buffer size.}

\item{cores}{Number of cores to use during paralllel processing. Defaults to 2.}

\item{sg}{spatial grid created from \code{grid_res} function}

\item{df}{data frame object consisting of results from iterative reconstruction process}

\item{km}{grid cell resolution in km. one-sided length of grid cell, assumes desired grid cell is to be squared}

\item{study_site}{simple feature polygon object that encompasses the entire study site}

\item{epsg}{coordinate reference system specification for transformation}

\item{sts_pts}{a simple feature (multi)point object representing receiver locations with detection range buffer}

\item{depth_limit}{a depth value in meters that represents the new depth cut off of interest}
}
\description{
These functions are provided for backwards compatibility and will be removed
in a future release.
}
