% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth_data.R
\name{depth_data}
\alias{depth_data}
\title{Calculate mean depth}
\usage{
depth_data(HSgrid, depth)
}
\arguments{
\item{HSgrid}{list object of multiple grids}

\item{depth}{simple feature point object with \code{"altitude"} attribute data}
}
\value{
A simple feature multipolygon object with mean depth per grid cell
}
\description{
This function allows you to calculate mean depth per grid cell.
}
\examples{
depth <- readr::read_csv(file = "Depth_Data.csv") \%>\%
sf::st_as_sf(coords = c("longitude", "latitude"), crs = 4269) 
depth <- sf::st_filter(depth, site_depth) \%>\% 
mutate(uid = seq_along(geometry)) HS_100km_grid <- depth_data(HS_100km_grid, depth)
}
