% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fo_study_site}
\alias{fo_study_site}
\title{American Atlantic coastal shelf to the 500m isobath from Florida to Cape Code, MA}
\format{
An sf object with 1 row and two columns:
\describe{
\item{poly_uid}{Polygon ID}
\item{geometry}{Well-known text POLYGON of the American Atlantic coastal shelf
in WGS84/Pseudo-Mercator, EPSG:3857}
}
}
\usage{
fo_study_site
}
\description{
American Atlantic coastal shelf to the 500m isobath from Florida to Cape Code, MA
}
\keyword{datasets}
