% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_fts.R
\name{gd_fts}
\alias{gd_fts}
\title{Good fits summary statistics}
\usage{
gd_fts(df)
}
\arguments{
\item{df}{data frame object consisting of results from iterative reconstruction process}
}
\value{
a data frame object with summary statistics from grid cells that contained good fits
}
\description{
This function summarizes minimum, maximum, and mean statistics for depth (mean_depth), distance from shore (d_shore), receiver count (count), receiver presence/absence (p_a), and the percentage of grid cells missing receivers for all of the good fits (g_fit).
}
\examples{
gd_fit_char <- lapply(res, gd_fts) \%>\%
data.table::rbindlist(., idcol = 'resolution') \%>\%
 left_join(tibble(resolution = 1:4,
                  res_name = c("100km", "50km", "25km", "10km")),
          by = "resolution") \%>\%
 mutate(res_name = ordered(res_name, levels = c("100km", "50km", "25km", "10km")))
}
