% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_chr.R
\name{gps_chr}
\alias{gps_chr}
\title{Gaps characteristics}
\usage{
gps_chr(df)
}
\arguments{
\item{df}{data frame object consisting of results from iterative reconstruction process}
}
\value{
a data frame object with summary statistics that provide insight into the locations of the gaps in the network receiver array that were closed by the reconstruction process
}
\description{
Gaps characteristics
}
\examples{
#' res <- load(system.file("extdata", "res.Rda", package = "sporeg"))
clsd_gp_chars <- lapply(res, gps_chr) \%>\%
data.table::rbindlist(., idcol = 'resolution') \%>\%
  left_join(tibble(resolution = 1:4,
                   res_name = c("100km", "50km", "25km", "10km")),
            by = "resolution") \%>\%
            mutate(res_name = ordered(res_name, levels = c("100km", "50km", "25km", "10km")))
}
