% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_fld.R
\name{gps_fld}
\alias{gps_fld}
\title{Gaps filled}
\usage{
gps_fld(df)
}
\arguments{
\item{df}{data frame object consisting of results from iterative reconstruction process}
}
\value{
a data frame object with a percentage of grid cells that contained a good fit out of those that lacked receivers
}
\description{
This function calculates how well the reconstructions closed gaps in the network receiver array
}
\examples{
# Apply gps_fld to list of grid resolutions

library(sporeg)
library(dplyr)
library(sf)
library(data.table)
load(system.file("extdata", "res.Rda", package = "sporeg"))

clsd_gps <- lapply(res, gps_fld) \%>\%
data.table::rbindlist(., idcol = 'resolution') \%>\%
 left_join(tibble(resolution = 1:4,
                   res_name = c("100km", "50km", "25km", "10km")),
           by = "resolution") \%>\%
 dplyr::mutate(res_name = ordered(res_name, levels = c("100km", "50km", "25km", "10km")))
}
