% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_line.R
\name{make_line}
\alias{make_line}
\title{Make a line}
\usage{
make_line(start_x, start_y, end_x, end_y)
}
\arguments{
\item{start_x}{longitudinal sf coordinate object from the desired start point}

\item{start_y}{latitudinal sf coordinate object from the desired start point}

\item{end_x}{longitudinal sf coordinate object from the desired end point}

\item{end_y}{latitudinal sf coordinate object from the desired end point}
}
\value{
A simple feature geometry object or simple feature linestring object
}
\description{
This functions creates a line by connecting start and end points.
}
\examples{
library(sporeg)
library(dplyr)
library(sf)
library(purrr)
library(tidyr)

at_lines <- at_dly_locs \%>\%
 dplyr::group_by(ID, time) \%>\%
 sf::st_transform(3857) \%>\%
 dplyr::summarise(pt = sf::st_combine(geometry)) \%>\%
 sf::st_centroid() \%>\%
 dplyr::mutate(lat = sf::st_coordinates(pt)[,2],
        lon = sf::st_coordinates(pt)[,1]) \%>\%
 dplyr::arrange(ID, time) \%>\% #Order data for making lines
 dplyr::mutate(start_x = lon, start_y = lat,
 end_x = dplyr::lead(lon), end_y = dplyr::lead(lat)) \%>\%
 sf::st_as_sf(coords = c("lon", "lat"), crs = 3857) \%>\%
 dplyr::filter(!is.na(end_y)) \%>\%
 tidyr::nest() \%>\%
 dplyr::mutate(
   data = purrr::map(data,
                     ~ dplyr::mutate(.x,
                              x = purrr::pmap(.l = list(start_x, start_y, end_x, end_y),
                                              .f = make_line))))
}
