% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_line.R
\name{make_line}
\alias{make_line}
\title{Make a line}
\usage{
make_line(start_x, start_y, end_x, end_y)
}
\arguments{
\item{start_x}{longitudinal sf coordinate object from the desired start point}

\item{start_y}{latitudinal sf coordinate object from the desired start point}

\item{end_x}{longitudinal sf coordinate object from the desired end point}

\item{end_y}{latitudinal sf coordinate object from the desired end point}
}
\value{
A simple feature geometry object or simple feature linestring object
}
\description{
This functions creates a line by connecting start and end points.
}
\examples{
make_line(Longitude, Latitude, lead(Longitude), lead(Latitude))
make_line(sf::st_coordinates(point_geom)[,1], sf::st_coordinates(point_geom)[,2],
lead(sf::st_coordinates(point_geom)[,1]), lead(sf::st_coordinates(point_geom)[,2]))
}
