% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_line.R
\name{make_line}
\alias{make_line}
\title{Make a line}
\usage{
make_line(start_x, start_y, end_x, end_y)
}
\arguments{
\item{start_x}{longitudinal sf coordinate object from the desired start point}

\item{start_y}{latitudinal sf coordinate object from the desired start point}

\item{end_x}{longitudinal sf coordinate object from the desired end point}

\item{end_y}{latitudinal sf coordinate object from the desired end point}
}
\value{
A simple feature geometry object or simple feature linestring object
}
\description{
This functions creates a line by connecting start and end points.
}
\examples{
load(system.file("extdata", "at_dly_locs.Rda", package = "sporeg"))
at_lines <- at_dly_locs \%>\%
group_by(ID, time) \%>\%
sf::st_transform(3857) \%>\%
summarise(pt = sf::st_combine(geometry)) \%>\%
sf::st_centroid() \%>\%
mutate(lat = sf::st_coordinates(pt)[,2],
lon = sf::st_coordinates(pt)[,1]) \%>\%
arrange(ID, time) \%>\%
mutate(start_x = lon, start_y = lat, end_x = lead(lon), end_y = lead(lat)) \%>\% #Prep data for making lines - ensures proper order
sf::st_as_sf(coords = c("lon", "lat"), crs = 3857) \%>\%
filter(!is.na(end_y)) \%>\%
tidyr::nest() \%>\%
mutate(
data = purrr::map(data,
~ mutate(.x,
x = purrr::pmap(.l = list(start_x, start_y, end_x, end_y),
.f = make_line))))
}
