% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powr.R
\name{powr}
\alias{powr}
\title{Run a power analysis}
\usage{
powr(output, sig.level, power, delta, n)
}
\arguments{
\item{output}{the resulting data frame from the iterative methods process}

\item{sig.level}{numeric. the desired level of significance to achieve}

\item{power}{numeric. the desired level of power to achieve}

\item{delta}{numeric. the desired effect size to achieve}

\item{n}{integer. the number of replicates/sample size. should be assigned \code{NULL} if desiring sample size}
}
\value{
A data frame object
}
\description{
This is a wrapper function for running a power analysis on iterative simulation and reconstruction methods process.
}
\examples{
# Use powr wrapper function on example results

library(sporeg)
library(dplyr)
library(data.table)

results <- lapply(results, data.table::rbindlist, idcol = 'resolution')
results <- data.table::rbindlist(results, idcol = 'iteration') \%>\%
  dplyr::left_join(dplyr::tibble(resolution = 1:4,
    res_name = c("100km", "50km", "25km", "10km")),
    by = "resolution")

df_var <- zero_var(results)
pow_stat <- df_var \%>\%
  dplyr::group_by(res_name, gid) \%>\%
  dplyr::summarise(sd = sd(dif)) \%>\%
  dplyr::ungroup() \%>\%
  dplyr::group_by(res_name)

res <- pow_stat \%>\% dplyr::filter(res_name == "100km")

anims <- 30
power <- 0.8
delta <- anims*0.01 # a delta within 1\% of the total number of animals
sig.level <- 0.95
n <- NULL

pwr_100km <- powr(res, sig.level, power, delta)
print("Grid cell resolution: 100 km x 100 km")
pwr_100km
}
