% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powr.R
\name{powr}
\alias{powr}
\title{Run a power analysis}
\usage{
powr(output, sig.level, power, delta, n)
}
\arguments{
\item{output}{the resulting data frame from the iterative methods process}

\item{sig.level}{numeric. the desired level of significance to achieve}

\item{power}{numeric. the desired level of power to achieve}

\item{delta}{numeric. the desired effect size to achieve}

\item{n}{integer. the number of replicates/sample size. should be assigned \code{NULL} if desiring sample size}
}
\value{
A data frame object
}
\description{
This is a wrapper function for running a power analysis on iterative simulation and reconstruction methods process.
}
\examples{
df_var <- zero_var(df)
pow_stat <- df_var \%>\%
group_by(res_name, gid) \%>\%
summarise(sd = sd(dif)) \%>\%
ungroup() \%>\%
group_by(res_name)

res <- pow_stat \%>\% dplyr::filter(res_name == "100km")
pwr_100km <- powr(res, sig.level, power, delta)
print("Grid cell resolution: 100 km x 100 km")
pwr_100km
}
