% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcv_chr.R
\name{rcv_chr}
\alias{rcv_chr}
\title{Receiver characteristics}
\usage{
rcv_chr(df)
}
\arguments{
\item{df}{data frame object consisting of results from iterative reconstruction process}
}
\value{
a data frame object with summary statistics that provide insight into the locations of the receivers in the network receiver array
}
\description{
Receiver characteristics
}
\examples{
res <- load(system.file("extdata", "res.Rda", package = "sporeg"))
rcv_chars <- lapply(res, rcv_chr) \%>\%
data.table::rbindlist(., idcol = 'resolution') \%>\%
left_join(tibble(resolution = 1:4,
res_name = c("100km", "50km", "25km", "10km")),
            by = "resolution") \%>\%
              mutate(res_name = ordered(res_name, levels = c("100km", "50km", "25km", "10km")))
}
