% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul_trks.R
\name{simul_trks}
\alias{simul_trks}
\title{Simulate tracks}
\usage{
simul_trks(
  anims,
  study_site,
  theta,
  vmin,
  vmax,
  rel_site,
  crs,
  n_days,
  initHeading
)
}
\arguments{
\item{anims}{integer. quantity of desired animals to simulate}

\item{study_site}{simple feature polygon object that encompasses area where tracks are allowed to be simulated.}

\item{theta}{argument from \code{glatos::crw_in_polygon} function}

\item{vmin}{numeric. minimum velocity from which to sample step length}

\item{vmax}{numeric. maximum velocity from which to sample step length}

\item{rel_site}{simple feature polygon object in which simulated animals are "released" or where simulated tracks begin}

\item{crs}{EPSG code for study_site polygon. Must be projected coordinate system}

\item{n_days}{integer. number of days that tracks should be simulated}

\item{initHeading}{argument from \code{glatos::crw_in_polygon} function}
}
\value{
simple feature (multi)linestring object that represents individual simulated tracks
}
\description{
This function simulates tracks inside a specified polygon.
}
\examples{
 df <- simul_trks(anims, study_site, theta, vmin, vmax, rel_site, n_days, initHeading)
}
