% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul_trks.R
\name{simul_trks}
\alias{simul_trks}
\title{Simulate tracks}
\usage{
simul_trks(
  anims,
  study_site,
  theta,
  vmin,
  vmax,
  rel_site,
  crs,
  n_days,
  initHeading
)
}
\arguments{
\item{anims}{integer. quantity of desired animals to simulate}

\item{study_site}{simple feature polygon object that encompasses area where tracks are allowed to be simulated.}

\item{theta}{argument from \code{glatos::crw_in_polygon} function}

\item{vmin}{numeric. minimum velocity from which to sample step length}

\item{vmax}{numeric. maximum velocity from which to sample step length}

\item{rel_site}{simple feature polygon object in which simulated animals are "released" or where simulated tracks begin}

\item{crs}{EPSG code for study_site polygon. Must be projected coordinate system}

\item{n_days}{integer. number of days that tracks should be simulated}

\item{initHeading}{argument from \code{glatos::crw_in_polygon} function}
}
\value{
simple feature (multi)linestring object that represents individual simulated tracks
}
\description{
This function simulates tracks inside a specified polygon.
}
\examples{
library(sporeg)
library(sf)
library(glatos)

load(system.file("extdata", "rel_site.Rda", package = "sporeg"))
load(system.file("extdata", "study_site.Rda", package = "sporeg"))

anims <- 30
yr <- 1
theta <- c(0, 1.74)
vmin <- 0.98
vmax <- 1.58
crs <- 3857
n_days <- 365*yr
initHeading <- 0

tracks <- simul_trks(anims, study_site, theta, vmin, vmax, rel_site, crs, n_days, initHeading)
}
