% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_rrt.R
\name{sub_rrt}
\alias{sub_rrt}
\title{Re-route tracks around a barrier}
\usage{
sub_rrt(track_data, CRS, barrier, vis_graph, buffer)
}
\arguments{
\item{track_data}{point location data with latitude and longitude information}

\item{CRS}{epsg code for desired coordinate system transformation}

\item{barrier}{polygon or multipolygon sf object}

\item{vis_graph}{vis_graph object created from \code{pathroutr::prt_vis_graph} function}

\item{buffer}{desired buffer size for re-routed tracks; generally should relate to range of receivers; distance should use the same units as the final coordinate system}
}
\value{
A simple feature (multi)polygon object
}
\description{
This function allows you to re-route demonstrative movement data that has been reconstructed using a movement model around a polygon barrier.
}
\examples{
library(sporeg)
library(dplyr)
library(pathroutr)
library(sf)
load(system.file("extdata", "subset.Rda", package = "sporeg"))
load(system.file("extdata", "atlcoast.Rda", package = "sporeg"))

CRS <- 3857
barrier <- atlcoast
vis_graph <- pathroutr::prt_visgraph(barrier)
buffer <- 650

tbuff650 <- sub_rrt(subset, CRS, barrier, vis_graph, buffer)

}
