% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_var.R
\name{zero_var}
\alias{zero_var}
\title{Remove zero variance}
\usage{
zero_var(df)
}
\arguments{
\item{df}{data frame object of iterative methods results}
}
\value{
a data frame object with grid IDs that did not exhibit zero variance
}
\description{
This function removes grid cells that lacked any variance.
}
\examples{
# Remove grid cells with zero variance
res <- lapply(sporeg::results, `[[`, 1) |>
  dplyr::bind_rows() |>
  dplyr::mutate(res_name = "100km") |>
  # Add 50 rows with no variance for demonstration purposes
  dplyr::bind_rows(
    data.frame(gid = rep(100, 50), dif = rep(100, 50), res_name = "100km")
  )

df_var <- zero_var(res)
nrow(res) > nrow(df_var)
}
