% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_var.R
\name{zero_var}
\alias{zero_var}
\title{Remove zero variance}
\usage{
zero_var(df)
}
\arguments{
\item{df}{data frame object of iterative methods results}
}
\value{
a data frame object with grid IDs that did not exhibit zero variance
}
\description{
This function removes grid cells that lacked any variance.
}
\examples{
# Bind the lists by grid cell resolution
df <- lapply(df, data.table::rbindlist, idcol = 'resolution')
# Bind the bound resolution by iteration
df <- data.table::rbindlist(df, idcol = 'iteration')
df <- df \%>\%
left_join(tibble(resolution = 1:4, res_name = c("100km", "50km", "25km", "10km")), by = "resolution") 
# Remove grid cells with zero variance
df_var <- zero_var(df) 
}
